moved messages around
